feat(WinningNumbers): 당첨 번호 클래스 구현 및 검증 로직 인터페이스화
당첨 번호 클래스인 WinningNumbers를 구현하였습니다. 구현하고 보니, Lotto 클래스와 검증 로직이 겹쳐, static 메서드를 가진 Validator를 구현할지, default 메서드를 가진 interface를 구현할지에 대해 오랜 시간 고민했습니다.
일단은 제공해주신 Lotto 클래스의 검증 메서드 구현 방식을 해치지 않고 싶었기 때문에, Interface를 선택하였습니다. 또한, 자기 자신이 들어온 값을 검증한다 는 의미를 퇴색시키고 싶지 않기 때문이기도 합니다.
