+++ docs/README.md
@@ -60,9 +60,10 @@
     - [X] 1,000원으로 나누어 떨어지는 금액만 입력할 수 있다.
 
 ### 당첨 번호
-- [ ] 6개의 숫자로 구성되어 있다.
-  - [ ] 중복된 숫자가 있으면 안된다.
-  - [ ] 당첨 번호에 포함되어있으면 안된다.
+- [X] 6개의 숫자로 구성되어 있다.
+  - [X] 중복된 숫자가 있으면 안된다.
+  - [X] 숫자 범위는 1~45이다.
+  - [X] 뽑은 숫자들을 정렬한다.
 
 ### 보너스 번호
 - [ ] 1개의 숫자로 구성되어 있다.
