+++ src/main/java/lotto/domain/lotto_numbers/LottoNumbersErrorMessages.java
@@ -1,0 +1,8 @@
+package lotto.domain.lotto_numbers;
+
+public class LottoNumbersErrorMessages {
+
+    static final String INVALID_SIZE_OF_LOTTO_NUMBERS = "로또 숫자는 6개만 입력할 수 있습니다.";
+    static final String NOT_ALLOW_DUPLICATED_LOTTO_NUMBER = "중복 숫자는 입력될 수 없습니다.";
+    static final String INVALID_RANGE_OF_LOTTO_NUMBER = "로또 번호는 1부터 45 사이의 숫자여야 합니다.";
+}
